# Faster English Porter's Stemmer