This gets rid of the extremely slow regex based implementation & introduces an all new implementation taken from [rust-stem](https://github.com/minhnhdo/rust-stem). I had to make a few changes to make it compatible with elasticlunr.

How fast is it exactly?

Old:
```
create_index            time:   [3.4562 ms 3.4664 ms 3.4838 ms]
```
New:
```
create_index            time:   [509.29 µs 511.17 µs 513.40 µs]                         
                        change: [-85.031% -84.680% -84.282%] (p = 0.00 < 0.05)
                        Performance has improved.
```

All tests are passing without any changes to them or the outputs so this is 100% compatible. I bet this can be further improved but it's a good start.