@thecodrr 2022-11-11
> Would anybody be bothered if it took 100ms to build the search index

No harm in having better performance if it's possible without any compromises. The stemmer was one main bottleneck which this PR aims to remove. Having better performance means you can do other more important stuff in the same amount of time. A few ms here and there quickly add up to create a very clunky & slow UX.

If we talk especially about index building then it doesn't matter if you make the Index only once but if you are doing it every few seconds (e.g. in case of generating docs) then it quickly starts to make a difference in your productivity. If we had incremental indexing that'd help alot in that use case but since the index is built from scratch every single time, it needs to be as fast as possible.

> We'd want to at least add the [license header from rust-stem](https://github.com/minhnhdo/rust-stem/blob/master/LICENSE) if we went this route.

Of course. I will make the necessary changes ASAP.